server/requirements.txt
Flask==1.1.2
python-dotenv==0.15.0
flask-sqlalchemy==2.4.4
flask-migrate==2.5.3
flask-cors==3.0.9
flask_mail==0.9.1
pytest==6.2.1
PyJWT==1.7.1
psycopg2-binary==2.8.6
google-api-python-client==1.12.8
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
